backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.7
flower==2.0.1
gunicorn==21.2.0
humanize==4.8.0
inflection==0.5.1
kombu==5.3.2
packaging==23.2
Pillow==10.0.1
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psutil==5.9.5
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.9
